Use real abi length
