+++ crates/forge-runner/src/lib.rs
@@ -101,6 +101,7 @@ pub fn run_for_test_case(
     args: Vec<ConcreteTypeLongId>,
     case: Arc<TestCaseWithResolvedConfig>,
     casm_program: Arc<AssembledProgramWithDebugInfo>,
+    code_size: usize,
     forge_config: Arc<ForgeConfig>,
     versioned_program_path: Arc<Utf8PathBuf>,
     send: Sender<()>,
@@ -110,6 +111,7 @@ pub fn run_for_test_case(
             let res = run_test(
                 case,
                 casm_program,
+                code_size,
                 forge_config.test_runner_config.clone(),
                 versioned_program_path,
                 send,
@@ -123,6 +125,7 @@ pub fn run_for_test_case(
                 args,
                 case,
                 casm_program,
+                code_size,
                 forge_config.test_runner_config.clone(),
                 versioned_program_path,
                 send,
@@ -154,6 +157,7 @@ fn run_with_fuzzing(
     args: Vec<ConcreteTypeLongId>,
     case: Arc<TestCaseWithResolvedConfig>,
     casm_program: Arc<AssembledProgramWithDebugInfo>,
+    code_size: usize,
     test_runner_config: Arc<TestRunnerConfig>,
     versioned_program_path: Arc<Utf8PathBuf>,
     send: Sender<()>,
@@ -187,6 +191,7 @@ fn run_with_fuzzing(
                 args,
                 case.clone(),
                 casm_program.clone(),
+                code_size,
                 test_runner_config.clone(),
                 versioned_program_path.clone(),
                 send.clone(),
