+++ crates/forge-runner/src/running.rs
@@ -19,7 +19,6 @@ use cairo_lang_sierra::extensions::segment_arena::SegmentArenaType;
 use cairo_lang_sierra::extensions::starknet::syscalls::SystemType;
 use cairo_lang_sierra::extensions::NamedType;
 use cairo_lang_sierra::ids::GenericTypeId;
-use cairo_lang_sierra::program::Program;
 use cairo_lang_utils::unordered_hash_set::UnorderedHashSet;
 use cairo_vm::vm::errors::cairo_run_errors::CairoRunError;
 use cairo_vm::vm::errors::vm_errors::VirtualMachineError;
@@ -44,7 +43,6 @@ use entry_code::create_entry_code;
 use hints::{hints_by_representation, hints_to_params};
 use runtime::starknet::context::{build_context, set_max_steps};
 use runtime::{ExtendedRuntime, StarknetRuntime};
-use starknet_api::contract_class::FELT_WIDTH;
 use starknet_types_core::felt::Felt;
 use std::cell::RefCell;
 use std::default::Default;
@@ -68,7 +66,7 @@ pub mod with_config;
 pub fn run_test(
     case: Arc<TestCaseWithResolvedConfig>,
     casm_program: Arc<AssembledProgramWithDebugInfo>,
-    sierra_program: Program,
+    code_size: usize,
     test_runner_config: Arc<TestRunnerConfig>,
     versioned_program_path: Arc<Utf8PathBuf>,
     send: Sender<()>,
@@ -84,6 +82,7 @@ pub fn run_test(
             vec![],
             &case,
             &casm_program,
+            code_size,
             &RuntimeConfig::from(&test_runner_config),
         );
 
@@ -108,6 +107,7 @@ pub(crate) fn run_fuzz_test(
     args: Vec<Felt>,
     case: Arc<TestCaseWithResolvedConfig>,
     casm_program: Arc<AssembledProgramWithDebugInfo>,
+    code_size: usize,
     test_runner_config: Arc<TestRunnerConfig>,
     versioned_program_path: Arc<Utf8PathBuf>,
     send: Sender<()>,
@@ -125,6 +125,7 @@ pub(crate) fn run_fuzz_test(
             args.clone(),
             &case,
             &casm_program,
+            code_size,
             &Arc::new(RuntimeConfig::from(&test_runner_config)),
         );
 
@@ -158,6 +159,7 @@ pub fn run_test_case(
     args: Vec<Felt>,
     case: &TestCaseWithResolvedConfig,
     casm_program: &AssembledProgramWithDebugInfo,
+    code_size: usize,
     runtime_config: &RuntimeConfig,
 ) -> Result<RunResultWithInfo> {
     assert!(args.is_empty(), "Tests with args not supported currently");
@@ -300,9 +302,7 @@ pub fn run_test_case(
         .calldata
         .clone();
     let used_resources = get_all_used_resources(forge_runtime, &transaction_context);
-    // TODO: Calcualate or retrieve `code_size`
-    // Ref: https://github.com/starkware-libs/sequencer/blob/7319f200db1df692be89245c43bfaf8af595df9d/crates/starknet_api/src/contract_class.rs#L159
-    let code_size = calculate_code_size(0, 0, 0);
+
     let gas = calculate_used_gas(
         &transaction_context,
         &mut cached_state,
@@ -329,15 +329,6 @@ pub fn run_test_case(
         call_trace: call_trace_ref,
         encountered_errors,
     })
-}
-
-// Ref: https://github.com/starkware-libs/sequencer/blob/7319f200db1df692be89245c43bfaf8af595df9d/crates/starknet_api/src/contract_class.rs#L159
-fn calculate_code_size(
-    bytecode_length: usize,
-    sierra_program_length: usize,
-    abi_length: usize,
-) -> usize {
-    (bytecode_length + sierra_program_length) * FELT_WIDTH + abi_length
 }
 
 // FIXME get rid of copied code
